code/embedding-search/embedding-func/requirements.txt
google-cloud-storage
google-cloud-aiplatform
vertexai